+++ griptape/drivers/prompt/ollama_prompt_driver.py
@@ -1,7 +1,7 @@
 from __future__ import annotations
 
+import json
 import logging
-from collections.abc import Iterator
 from typing import TYPE_CHECKING, Any, Optional
 
 from attrs import Factory, define, field
@@ -31,6 +31,8 @@ from griptape.utils.decorators import lazy_property
 logger = logging.getLogger(Defaults.logging_config.logger_name)
 
 if TYPE_CHECKING:
+    from collections.abc import Iterator
+
     from ollama import ChatResponse, Client
 
     from griptape.tokenizers.base_tokenizer import BaseTokenizer
@@ -92,14 +94,18 @@ class OllamaPromptDriver(BasePromptDriver):
     def try_stream(self, prompt_stack: PromptStack) -> Iterator[DeltaMessage]:
         params = {**self._base_params(prompt_stack), "stream": True}
         logger.debug(params)
-        stream = self.client.chat(**params)
+        stream: Iterator = self.client.chat(**params)
 
-        if isinstance(stream, Iterator):
-            for chunk in stream:
-                logger.debug(chunk)
-                yield DeltaMessage(content=self.__to_prompt_stack_delta_message_content(chunk))
-        else:
-            raise Exception("invalid model response")
+        tool_index = 0
+        for chunk in stream:
+            logger.debug(chunk)
+            message_content = self.__to_prompt_stack_delta_message_content(chunk)
+            # Ollama provides multiple Tool calls as separate chunks but with no index to differentiate them.
+            # So we must keep track of the index ourselves.
+            if isinstance(message_content, ActionCallDeltaMessageContent):
+                message_content.index = tool_index
+                tool_index += 1
+            yield DeltaMessage(content=message_content)
 
     def _base_params(self, prompt_stack: PromptStack) -> dict:
         messages = self._prompt_stack_to_messages(prompt_stack)
@@ -115,7 +121,7 @@ class OllamaPromptDriver(BasePromptDriver):
             params["format"] = prompt_stack.output_schema.json_schema("Output")
 
         # Tool calling is only supported when not streaming
-        if prompt_stack.tools and self.use_native_tools and not self.stream:
+        if prompt_stack.tools and self.use_native_tools:
             params["tools"] = self.__to_ollama_tools(prompt_stack.tools)
 
         return params
@@ -243,27 +249,19 @@ class OllamaPromptDriver(BasePromptDriver):
         message = content_delta["message"]
         if "content" in message and message["content"]:
             return TextDeltaMessageContent(message["content"])
-        elif "tool_calls" in message:
+        elif "tool_calls" in message and len(message["tool_calls"]):
             tool_calls = message["tool_calls"]
 
-            if len(tool_calls) == 1:
-                tool_call = tool_calls[0]
-                index = tool_call.index
+            # Ollama doesn't _really_ support Tool streaming. They provide the full tool call at once.
+            # Multiple, parallel, Tool calls are provided as multiple content deltas.
+            # Tracking here: https://github.com/ollama/ollama/issues/7886
+            tool_call = tool_calls[0]
 
-                if tool_call.function is not None:
-                    # Tool call delta either contains the function header or the partial input.
-                    if tool_call.id is not None and tool_call.function.name is not None:
-                        return ActionCallDeltaMessageContent(
-                            index=index,
-                            tag=tool_call.id,
-                            name=ToolAction.from_native_tool_name(tool_call.function.name)[0],
-                            path=ToolAction.from_native_tool_name(tool_call.function.name)[1],
-                        )
-                    else:
-                        return ActionCallDeltaMessageContent(index=index, partial_input=tool_call.function.arguments)
-                else:
-                    raise ValueError(f"Unsupported tool call delta: {tool_call}")
-            else:
-                raise ValueError(f"Unsupported tool call delta length: {len(tool_calls)}")
+            return ActionCallDeltaMessageContent(
+                tag=tool_call["function"]["name"],
+                name=ToolAction.from_native_tool_name(tool_call["function"]["name"])[0],
+                path=ToolAction.from_native_tool_name(tool_call["function"]["name"])[1],
+                partial_input=json.dumps(tool_call["function"]["arguments"]),
+            )
         else:
             return TextDeltaMessageContent("")