+++ CHANGELOG.md
@@ -22,6 +22,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+## Added
+- Tool streaming support to `OllamaPromptDriver`.
+
 ## [1.2.0] - 2025-01-21
 
 ### Added